Enforce abstraction boundaries: extract agent creator flow and decouple auth notifications
Move the five PocketFlow Node classes and flow factory out of the VS Code
command layer into src/agent/implementations/flows/agentCreator/, which is
the correct home for all Node implementations. The nodes now receive all
host-specific operations (input prompts, directory resolution, file opening)
through an injected AgentCreatorUI interface. The command file becomes a thin
wiring layer that provides the VS Code implementations of that interface.

Add an AuthNotifier port to SupabaseAuthProvider, injected at construction
from extension.ts, replacing seven direct vscode.window.show* calls inside
the auth provider's business logic. The handleInvalidSession sign-in prompt
is moved to the notifier so the provider no longer owns its own UI.
